Created separate files for both db tables
